requirements.txt
discord.py
timeago
jishaku
psutil
numpy
parsedatetime
aiohttp
asyncpg
asyncio
aiohttp
googletrans
praw
wand
humanize
dropbox
pathlib
asyncio
statcord.py
python-dateutil
discordlists.py